Use rsync to exclude git files in helm copy (#790)
When make images was re-run (like after a build failure somewhere in
the process), this step would try to replace existing files in the
destination .git directory that were read-only, and so it would fail.

Rather than using cp's -f flag or similar, I figured, why not just
exclude .git entirely? We don't want that going into the helm chart
anyway.

If rsync isn't available in the builder base, this could be a problem
I guess, but I think rsync is pretty standard, and the flags I'm using
have been around for a long time.