Create helm chart repos
The create-ecr-repos Makefile target is extended to create the repos
of helm charts as well. In many cases the helm chart will end up in
one of the repos that's already created for a container image, but in
other cases, they're different repos, and this code will handle that
situation.