retries docker pull and use builder base tag (#833)
Mitali ran into a couple issues during the staging build.

We got hit by the random tls handshake timeout when trying to pull
the qemu image and since the builder base had been updated with new
golang versions, the source controller build pulled the latest
builder base and the checksums did not match. Abhay had added the
builder base tag env var and config file, for code build and prow
respectively, but we never implemented using them. With this change
the source controller cgo build should always pull the same builder
base that the current build is running in, instead of the latest.