helm/build: Read manifest digest from the local docker image
Previously, the manifest digest was being read from the image in its
repository. The problem with that, is that if you're building locally for the
first time, then that image hasn't been pushed to the repository yet, hence
skopeo would fail to read the manifest, or its digest. Furthermore, if the
image did exist in the repository, but it differed from the one you just built
locally, then the manifest digest was just plain incorrect.

To read the manifest from the local image, the helm Makefile rules are
modified to more closely mimic other project-specific pattern rules, like
%/images/amd64. This provides a way to pass in the project name, and from it
the path to the local tarball, from which the manifest digest is read.

I also added a more explicit check for the number of arguments passed to
helm_require.sh, because I've been bitten more than once by one of those
arguments being an empty string, which just causes the available arguments to
shift down, leaving HELM_IMAGE_LIST empty, and then the script gives strange
error messages. I might look into whether quoting the arguments to
helm_require.sh can also alleviate that problem.