Add Prometheus project (#1505)
* Add Prometheus project

* Update Prometheus project

Instead of embedding the web ui as a patch, we will build them through make targets first.

* Update project Prometheus makefile

This commit mainly makes two changes:
(1) re-write web UI build targets to reference a generated file instead of using PHONY target. This is to avoid re-building of the UI every time.
(2) add in a node 16 install step since presubmit uses node version 15, which is not compatible with this project.

* Update project Prometheus CHECKSUMS

* Add prometheus version badge and build status

* Update Prometheus Makefile and Dockerfile

Re-write make targets to reduce number of re-runs; update Dockerfile to use the minimal base image.

* Update UPSTREAM_PROJECTS