Update for Vlan support in Tinkerbell (#1529)
* Update Tinkerbell Boots version:

This new version provides VLAN tagging support.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Add CAPT patch for updating tink dependency:

The updated tink version has the support needed
for enabling VLAN tagging functionality.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Add patch for enabling vlan interface creation

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Update Tinkerbell Hook commit:

This commit has the functionality for VLAN tagging
and removes the need to do any patching in this repo.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Fix Hook build issues:

Update Hook's makefile to play nice with
build-tools container image.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Update hardware crd spec in tinkerbell helm chart:

This adds the vlan id field needed for trunk vlan support.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Add License copy back in:

It was commented out for testing.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Add hardware spec CAPI labels back.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>