Update for enabling tagged VLAN support in Tinkerbell stack. (#1762)
* Update for enabling tagged VLAN support in Tinkerbell stack.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Use tag instead of version:

This is needed because of an idiosyncrasy of Hook.
The top of tree commit always points to the git tag
of "latest". So our tooling here will find that this
commit hash has an associated git tag. Once the top of
tree commit in Hook changes (i.e. a PR is merged) the
"latest" git tag will not point to this commit has anymore
and we will need to update the back to "commit" instead of
"tag".

Hook tracking issue: https://github.com/tinkerbell/hook/issues/157

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

* Bump tinkerbell chart version

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>