Fix unschedulable Tinkerbell stack when no worker nodes (#2624)
* Fix unschedulable Tinkerbell stack when no worker nodes

If all worker nodes elgible to run Tinkerbell stack workloads go offline
the cluster and Workload clusters cannot recover failed nodes. This
change allows the Tinkerbell stack to run on control plane nodes but
prefer Worker nodes at scheduling time.

* Increase Tinkerbell chart patch version

* Update Tinkerbell chart tags

* Tolerate control plane taints in Tinkerbells KubeVIP

KubeVIP needs to run on control planes so that when the Tinkerbell stack
floats to a control plane node (with Envoy) the VIP can be hosted from
that node and accommodate the Local externalTrafficPolicy.